Added tasks for DHT11 sensor, and added coordinator task
Installed DHT sensor library by Adafruit.

Added two task for DHT11 sensor: one for the temperature and the other for the humidty as there are two different readings functions for the two sensor readings. These tasks send a message  to the message queue of the coordinator task. (explained in more detail below)

The coordinator task has been created and the relative message queue has been created and used to receive a struct containing the sensor id(implemented as an enum) and the sensor reading (as a float).