ADDED: Task reader and task actuator for the lights, file .cpp and .h for grouping together utility functions.
Added task reader that reads the amount of ambient light (lux), the actual reading has not been implemented yet.

Added task actuator for turning on and off the lights, two threshold have been used (one for the activation and the other for the deactivation of the lights).
The activation is performed if the amount of ambient light is below a threshold, while the deactivation is performed only if the lights are already on and the amount of ambient light is above another threshold.

Added debug option for printing coordinator queue usage (number of messages waiting ecc...).

Created file .cpp and .h for containing all the neccessary utility functions, at the moment it contains functions for printing info about stack and queue utilization.
All stack memory utilization debug prints are substituted by a call to the relative utility function.

Moved initializations (pin, sensors) at the beginning of the function setup, in order to make sure that those are executed before any task is created.