FIXED: lights activation logic
Now the lights will be activated only if they are off and the amount of ambient light is below a threshold.
They will be deactivated only by the coordinator (and no more by themselves after a delay) when the amount of light is above another threshold.

Optimized if statement for the light deactivation by putting "else if" instead of "if".