ADDED: Preferences system, connection phase at setup, protocol for updating threshold via mqtt, stack usage prints, irrigator minimum actuation duration UPDATED: coordinator behaviour. CORRECTED: misra c. Minor changes
ADDED:
Preferences system: now the esp32 saves the fetched settings from MQTT to the flash memory by using the "Preferences.h" library.
At startup the esp32 loads the settings present in flash memory.
The settings not present in flash will receive the default value defined in "SGreenHouseConfig.h".
The settings will then be updated using the values fetched from MQTT.
Settings in preferences are localized by using a key. Function "get_key_from_topic()" generate a unique key for a given topic name by returning its position inside the "MQTT_subscription_array"
(the position is converted to const char* by using the array "int_to_charpointer" that is inizialized by the function "init_get_key_from_topic()") .

Connection phase at setup: before any of the tasks start, a connection phase will be performed (by starting the TaskConnect task) which will try to perform a connection to the wifi network and to the mqtt broker for a certain (configurable) number of attempts.

Function MQTTFetchSubscriptions: wrapper of the method "mqtt.processPackets". Initializes the preferences for writing purposes.

Protocol for updating threshold with values fetched from mqtt. The logic of the protocol is implemented by functions which names start with "setThreshold",
the difference between these functions is the data type of the threshold that they have to update.
Functions setMinThreshold and setMaxThreshold are wrappers for the aforementioned core functions and they take as parameter a "struct threshold_setting_manage*"
that contains pointers to variables that are important for the proper functioning of the protocol (for example variables that contain the value of a pending threshold value).
There exist different instances of the struct threshold_setting_manage (one for each type of threshold: one for soil thresholds, one for air thresholds ecc...)
General behaviour:
	When a value of a threshold (new_thr) is fetched, if new_thr is valid (meaning that: supposing new_thr is a minimum threshold, then it's valid it's less than the respective max), then the respective threshold setting is updated.
	If new_thr is not valid, it is put in a pending state, from now on if a new other threshold (if new_thr is min then the other threshold is a max) is fetched, it is accepted only if it's valid compared to the other pending threshold, if it's not he too is put in a pending state.
	If a new_thr is valid compared to the other pending threshold then both are accepted.

Minimum valid irrigator actuation duration: if the value fetched from mqtt for the irrigator actuation duration is less than the permitted value (defined in "SGreenHouseConfig.h"),
then it is not accepted, is resetted to the old valid value and the old valid value is published to the subscribe topic of the irrigator actuation duration.

Stack usage prints: now every task prints its stack usage.

UPDATED:
Coordinator behaviour: now the coordinator task reads a number of "Amount_of_sensor_ids" of sensor reading values.

CORRECTED + MINOR CHANGES
Misra C 2012: corrections for required rules, except rules 8.2 and 18.8: they seem to be false positives from the tool.

Minor changes: added two additional attributes (the default setting value and the pointer to the setting variable) to the "MQTT_subscription" struct, variable names refactoring, vTaskDelay to vTaskDelayUntil for sensor reading tasks.